plugins/docker: unmarshal docker configs carefully
Fixes https://github.com/coreos/bugs/issues/2077

We've seen in practice that these configs actually do sometimes not have
Env, and those type assertions were scary.

In addition, the fact that the failure mode was to "fall through" and
return a nil value is also scary.

This commit adds much more explicit unmarshal logic around getting the
config contents in the docker plugin.

It's non-ideal that "Contents" can't produce an error, but reworking the
architecture of plugins to account for errors is a more significant
change, so returning "dummy" error values seems like the pragmatic
choice.